requirements.txt
# Core Requirements
flask==2.3.2
waitress==2.1.2
python-dotenv==1.0.0

# Computer Vision - Pinned versions
opencv-python==4.6.0.66
numpy==1.24.4  # Updated to satisfy albumentations

# YOLOv8 with compatible dependencies
ultralytics==8.0.0
torch==2.0.1
torchvision==0.15.2

# OCR Processing
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pytesseract==0.3.10
pillow==9.5.0

# Database
mysql-connector-python==8.0.33
sqlalchemy==2.0.15

# Utility Packages
werkzeug==2.3.6
python-dateutil==2.8.2
requests==2.31.0

# Explicitly add image augmentation libraries
albumentations==2.0.6
albucore==0.0.24